requirements.txt
streamlit
feedparser
google-genai
python-dotenv
